Migrate <radio-preference> to vuejs
Due to the requirement of a default selection, the base won't update
 the value if nothing has been set yet.
The default value will be kept in tact that way.

It also circumvents the issue of hooks not being "inherited" like in real classes.
Vue.js implements a type of subclassing where "super" is **always** called.
It's overridable, but it then applies to all components (merge strategies).
The option was taken **not** to mess with that.